requirements.txt
asgiref==3.8.1
boto3==1.35.0
botocore==1.35.0
Django==5.1
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
jmespath==1.0.1
pillow==10.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
